refactor(spec-verify): split helpers and reduce complexity to satisfy quality gate (GH-304)
spec-verify.js grew past the 400-line and per-function complexity caps.
Split into spec-verify-helpers.js (path/glob/comment-strip state machines)
and spec-verify-checkers.js (FILE_EXISTS / GREP / TEST_COUNT / REUSES
checkers + dispatch). spec-verify.js stays the CLI entry. Behavior is
byte-identical — all 46 spec-verify tests pass after the extraction.

Staged: spec-verify.js (slimmed to CLI+parsing), new spec-verify-helpers.js, new spec-verify-checkers.js.
